+++ src/traci/gui/MultiDrawArea.java
@@ -30,7 +30,7 @@ public class MultiDrawArea extends AbstractDrawArea implements DrawArea
     }
     
     @Override
-    public void draw(final int x, final int y, final Color color)
+    public void draw(final long x, final long y, final Color color)
     {
         for (final DrawArea area : areas)
         {
