+++ src/traci/gui/NullDrawArea.java
@@ -10,7 +10,7 @@ public class NullDrawArea extends AbstractDrawArea implements DrawArea
     }
     
     @Override
-    public void draw(final int x, final int y, final Color color)
+    public void draw(final long x, final long y, final Color color)
     {
     }
 }
